+++ booknetwork/routes.py
@@ -1,0 +1,6 @@
+from flask import Flask, render_template, url_for, flash, jsonify, request, session, redirect
+from booknetwork import app, db
+
+@app.route("/")
+def index():
+    return render_template('index.html')
